backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
reportlab
python-multipart
pydantic
python-dotenv
google-generativeai
